divide code
